+++ ESPixelStick/src/output/OutputRelay.hpp
@@ -58,13 +58,13 @@ public:
     virtual ~c_OutputRelay ();
 
     // functions to be provided by the derived class
-    void Begin ();                                         ///< set up the operating environment based on the current config (or defaults)
-    bool SetConfig (ArduinoJson::JsonObject & jsonConfig); ///< Set a new config in the driver
-    void GetConfig (ArduinoJson::JsonObject & jsonConfig); ///< Get the current config used by the driver
-    void Render ();                                        ///< Call from loop(),  renders output data
-    void GetDriverName (String& sDriverName);
-    void GetStatus (ArduinoJson::JsonObject & jsonStatus) { c_OutputCommon::GetStatus (jsonStatus); }
-    uint16_t GetNumChannelsNeeded () { return Num_Channels; }
+    void   Begin ();                                         ///< set up the operating environment based on the current config (or defaults)
+    bool   SetConfig (ArduinoJson::JsonObject & jsonConfig); ///< Set a new config in the driver
+    void   GetConfig (ArduinoJson::JsonObject & jsonConfig); ///< Get the current config used by the driver
+    void   Render ();                                        ///< Call from loop(),  renders output data
+    void   GetDriverName (String& sDriverName);
+    void   GetStatus (ArduinoJson::JsonObject & jsonStatus) { c_OutputCommon::GetStatus (jsonStatus); }
+    size_t GetNumChannelsNeeded () { return Num_Channels; }
 
 private:
 #   define OM_RELAY_CHANNEL_LIMIT           8
